+++ packages/sdk/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.6.3
 - Add `101badges` Adaptor.
+- Update `celo`, `quadrata` Adaptor.
+- Update Dependencies.
 
 ## 0.6.2
 - Improve `rss3`, `layer3`, `tokenblacklist` Adaptors.